feat: implement invoice and subscription management components
