+++ .idea/dictionaries/joao_gustavo.xml
@@ -1,7 +1,11 @@
 <component name="ProjectDictionaryState">
   <dictionary name="joao.gustavo">
     <words>
+      <w>adquirir</w>
       <w>alterações</w>
+      <w>assinar</w>
+      <w>assinatura</w>
+      <w>atividades</w>
       <w>ativo</w>
       <w>atual</w>
       <w>atualizadas</w>
@@ -10,13 +14,16 @@
       <w>cancelar</w>
       <w>caractere</w>
       <w>caracteres</w>
+      <w>cobrança</w>
       <w>contratar</w>
       <w>criado</w>
       <w>deve</w>
+      <w>documentos</w>
       <w>encontrada</w>
       <w>erro</w>
       <w>fatura</w>
       <w>faturamento</w>
+      <w>finaliza</w>
       <w>gerado</w>
       <w>inativo</w>
       <w>incompleto</w>
@@ -29,7 +36,9 @@
       <w>ocultar</w>
       <w>pelo</w>
       <w>processar</w>
+      <w>próxima</w>
       <w>próximo</w>
+      <w>reativar</w>
       <w>salvar</w>
       <w>senha</w>
       <w>sucesso</w>
