+++ Policy Generator Requests/plans/stripe/cancel-subscription.bru
@@ -1,0 +1,17 @@
+meta {
+  name: cancel-subscription
+  type: http
+  seq: 3
+}
+
+post {
+  url: {{plans}}/create-checkout-session
+  body: json
+  auth: inherit
+}
+
+body:json {
+  {
+    "planId":"1296cf67-6cc5-4aa0-9b73-62a4780d3681"
+  }
+}
